requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.5.7
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-telegram-bot==13.12
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.6.3
tzlocal==5.0.1
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
black==23.3.0
cachetools==4.2.2
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
clickatell==0.1.2
cligj==0.7.2
coordinates==0.4.0
decorator==5.1.1
Fiona==1.9.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
isort==5.12.0
magic-filter==1.0.9
mock==5.0.2
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.7.0
pydantic==1.10.7
pyproj==3.5.0
python-box==7.0.1
python-dateutil==2.8.2
pytz==2023.3
ratelim==0.1.6
requests==2.28.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tomli==2.0.1
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
voluptuous==0.13.1

